Requirement #1512: Add new Study/Series Details screens to web GUI
Resolved the problem with toolbar buttons still not enabled in the  multiple server partition case. The problem was the script that listens to the study list component doesn't know or not aware of the fact that there are multiple instances of the toolbar buttons. The wrong buttons are attached to the wrong study list.

The solution is to convert the panel (SearchPanel) into a component with the toolbar and study list being the children. Each panel component has references to different instances of the buttons and study lists and therefore can link them up correctly. Note for unknown reasons, the panel component is instantiated only once while the buttons and the study list are instantiated each time a post back occurs. Therefore the panel component must listen to Sys.Application.Load event (fired after postback has completed) and re-link the buttons and study list.