Requirement: #1140
Updates to licenses. Updated the DICOM tool kit files that originated form the Colby Dillion tool kit to include Colby's new mDCM license.  Also updated the license screen on the server and the License.rtf file to include this License.rtf

Also included the OpenJPEG license in License.rtf and the server license screen..

Finally, made a small change to DicomPixelData so the GetFrame() method that outputs the PhotometricInterpretation is now in the base class.