Requirement: #1155 Design RIS automated testing infrastructure
Modified RisClient_dist.proj to copy ClearCanvas.Oto.Executable to the output directory.  This is a convenience, so that it can be run manually from the command-line in the same folder as the Desktop Executable, in order to share copies of the scripts between the IDE and the console app.