Ticket: #1546 - fix the inconsistent scaling of font sizes across the different pages.
This is my first stab at it. Try it out and give me feedback.

Approach: Forget about scaling relative to the "viewport" or width of the browser, since we want consistency of font-size which ever screen we might be looking at. Instead, make all the font sizing relative to the current default font size of the browser.

- Effectively get rid of scaleWidth(), by making it a no-op function.
- Changed the default font family to Arial, since Verdana has very large type. Used http://www.ampsoft.net/webdesign-l/WindowsMacFonts.html as a reference for browser fonts that are commonly available.
- Change em-type font-sizing to percentages
- Scale down the font size for the application to 80% of default size
- Tweak the patient name and MRN classes, by creating two new classes specifically for those headings (left current pageheading and pageheading2 untouched)