Requirement ##1396 : Add ability to set priority for WorkQueue entries
This is interim commit.

This commit includes the DB and UI changes. New buttons are added to the work queue search screen to either open the work queue details page or a dialog box to change the priority of the selected item. The work queue item priority and schedule date/time can be changed from the search screen or from the work queue details screen.

Note: .NET GridView control does keep track of the key for the selected item through SelectedIndex and SelectedDataKey properties. But their value change at different processing stages. Its value only truly reflects the user selection is between Loading and Data Binding stages. Once the new data is bound, the value may/may not be changed.

To reliably identify user's selection, the key for the selected row must be cached before the data is bound to the control again on a postback. The SelectedIndex must be re-evaluated after that, on DataBound event. Here is where WorkQueueItemCollection is used to locate the new row index based on the key.