Bug: #1902
Resolved problems with contention issues with the Study.xml file.  Added retry mechanism in the two places where the file is loaded such that if there is an exception thrown on the load, we retry opening the file after a random sleep between 5 and 50 ms.  We will retry up to 5 times before a failure occurs.  In testing, this seems to have resolved the problem.