Requirement: #1139
Basic working support for online compression.  This contains a number of changes/tweaks to get compression support working.  Fleshed out the code in the ServiceLock Item processor for compression, and fleshed out the WorkQueue compressor and rules engine support for setting compression parameters.  Still need to do a lot of testing, add support for setting compression parameters in the XML for rules, and do some GUI cleanup of some of the pull downs to show options better.

- Moved the action items for compression into the codec assemblies.Added new "Exempt" property of rules, so that we can specify a rule that exempts studies from having a rule applied.
- Added a number of new enumerated values for FilesystemQueue entries and rule types.
- Implemented first go arounds at actions for the various new rule types.
- Added a plugin mechanism for sample rules displayed int he web gui, so the rules are kept closer to the code.