Requirement: #2053 Refactor staff types Requirement: #2008 Add additional staff types
It was decided the application settings would be added to allow the supply filters for the Technologist 'Assign radiologist' and the Reporting 'Supervisor' lookup fields, thereby avoiding any hard-coded dependence on specific staff types existing.  This means however that if the default staff types are modified at deployment time, the application settings must be adjusted to ensure that it still behaves correctly.