Requirement: #1974 Search function should show broader results
The CreateAddendum operation needed to be modified to operate on a Procedure, rather than a PublicationStep, in order to be useful from search results (and this is probably a very common scenario - if the user is adding an addendum 6 months after the initial report, the only way they are going to find it is using the search).

CreateAddendum is the only reporting operation that can operate on a Procedure - everything else requires a specific step.