Defect #1705: Large header files should be handled properly and efficiently
This is interim commit. Only issues on the header streaming is fixed. The service can now send large header files without problems. Still need to investigate how to handle extremely large study in the study processor more efficiently. For large studies, it's a big problem having to load the entire study xml into memory when all we want to do is insert new sop into the xml. With the particular study used for testing, the headers for some sop are just so big due to private tags (4GB images with >3GB private data) that reloading the entire study xml will cause out of memory.

- Changed header streaming service to use basicHttpBinding instead of wsHttpBinding. There appears to be a limit around 40MB on the streamed data when using wsHttpBinding (in this binding, the data is actually buffered before being sent out).

- Added ReferenceID field in the header streaming service contract parameter, for tracking purpose.

- Updated test client.