Bug: #2203
Added fix for freezing up problem we were seeing in load testing on the server.  The problem is traced back to the lack of a DICOM ARTIM timeout after sending an association request in the library.  We were not receiving an Association Accept/Reject back from JDICOM, and the association was just stuck left open with no activity permanently, which caused a WorkQueue entry to be stuck.

Also with this change made a small update to how we display association information negotiated to reduce our log sizes.  We now just have one line in the log per presentation context.  The acccepted transfer syntax, presentation context info and abstract syntax are all on a single line.