Ticket: #2078
Updates to Find and Move Scus based on testing as part of the migration tool.  Cleaned up the Query Iods, and fixed some minor issues.  Need some more thought on if the QueryResultIod class should go away, and be replaced by the individual level Iods.  This needs some more thought.