Requirement: #2174 Order notes posted to me folder should be split
Users can now control which order notebox group folders are visible, and this is stored in a user setting.  By default a group is not visible, therefore no group folders are visible unless explicitly configured.  Configuration changes do not take affect until the home page is re-opened - ideally this will be corrected in a future commit.