Bug: #2284 Worklists with Group Subscribers added using the importer never contain items
The ticket description is misleading - the "Group Subscribers" part is a red herring.  The problem was caused by a bug in the importer that did not import the IncludeWorkingFacility flag on the Facilities filter, which would leave the worklist in an invalid state where it would not find any items.

The reason the workaround described in the ticket worked is because, by simply editing the worklist in the admin screen, the invalid facility filter would be corrected.