Requirement: #2065
Made change to the WorkQueue screen that we now display the first 60 characters of any failure description in the notes field for any non- dicom move entries in the workqueue.  The display was limited to 60 characters because the columns ended up being too close together on the screen when more characters were added.