Requirement: #1907
Changes in how connects fail with the DICOM tool kit and the ScuBase class.  We now detect and throw a specific NetworkException when a connect fails (we did not catch this before).  Also modified ScuBase to catch this condition, and added a new ScuOperationStatus status that sets the status to ConnectFailed.  Finally, updated the ImageServer auto-route processor to take this into account.