Bug: #1986 Decide what to do about folder Refresh
Refactored the entire auto-refresh mechanism.  Instead of each folder having its own built-in timer, there is now a single timer per FolderExplorerComponent, and it is the reponsibility of the FolderExplorerComponent to auto-refresh folders appropriately.  Note that most folders do not support auto-refresh - the only ones that do are the OrderNoteboxFolders.  The rationale for this design change is to give the FolderExplorerComponent the ability to suppress auto-refreshes when they are not needed (e.g when the folder explorer is not the active workspace).