#846: Filesystem Tier Support #847: Tier Migration Rules
This is interim commit.

It contains basic functional tier migration with limited testing. The implementation is different from what are specified in the tickets. Tiers are ordered based on their enumerated values (faster tiers are assigned with smaller enum values) and studies are migrated from higher to lower tiers when high watermark is hit. If no studies are eligible for migration, the systems will attempt to delete studies.

To prevent the queue from being filled up with too many migration in a row (hence, the system would not able to process new studies until all of migration are completed), migration work queue items are spread out according to the study sizes. That is, if a 5MB study takes approx 30 sec to move from one filesystem to another, next study migration will be scheduled approx 30 seconds after this study instead of immediately (<1 second). Other types of work queue can then be scheduled to occur in between this period.

Changes in this commit:

- Changed GetAll() in the enum classes to return List instead of IList. This allows us to use List.Sort() to sort filesystems based on tier order and size.
- Added tier migration work queue type and processor.
- Changed storage scp to store incoming studies in first available filesystem, sorted by tier.
- Changed filesystem delete service lock to attempt to schedule studies for migration first and if it can't be done, then delete studies. Also changed the code so that it no longer incrementally searches for studies to be delete in the future. The rationale is if users set up rules to keep the study on the tier for 1 year or delete the study after certain date, it may cause confusion (and make it harder to test) if the system removes the studies before that because of high watermark. It also defeats the purpose of allowing users to specify retention and delete rules if the systems will override them at the end. However, in the future we may still want to do incremental deletion to protect users from their stupidity.