Bug: #1986 Decide what to do about folder Refresh
Added icons that are displayed on the folders during the transient refresh state.  Note that the actual graphics are not necessarily what we want to use - the purpose here is more just a proof-of-concept, so we can see if it is beneficial to have the folders display an icon while they are refreshing.  Note also that the refreshing icons are actually animated gifs (the refresh symbol is supposed to rotate) but unfortunately WinForms does not play the animation.