Bug: #1986 Decide what to do about folder Refresh
Fixed a bunch of stuff related to maintaining the selection across refreshes - this code seems to have been broken for a long time now.  In any case, the new behaviour is somewhat different from the old behaviour: when the selected folder is refreshed, or when the selected folder changes, the FolderContentsComponent will attempt to re-select the same item that was previously selected prior to the refresh/folder change.  If the previous selection no longer exists in the updated items list, then the selection is cleared altogether.

This required some corrections to the logic in the TableView class, where it was comparing items by == rather than using the Equals method.