Ticket: #2078
During testing, found a number of cases where the underlying network / dicom code would shut down an association, but the ScuBase class would not be notified when this happened.  This was mainly related to ARTIM timeouts.  Updated the DICOM library so that we now call OnNetworkError whenever we have an internally detected failure and we're shutting down the connection.  This should cleanup cases where an Scu in ClearCanvas.DicomServices is frozen waiting for the child thread to complete, but the thread has already gone away.