Bug: #2324 Resubmit Protocol button is always disabled
In order to solve this problem, needed to refactor the IWorkflowService interface so that it is not a generic type. Although the generics provided some nice type-safety, they ultimately limit flexibility too much.  IWorkflowService.GetOperationEnablement can now operate on any object, and the WorkflowServiceBase implementation takes care of enforcing type safety at runtime.