Ticket: #1910 Ticket: #1911
Added framework for archiving and intro projects for an HSM archival mechanism.  Added new tables to the DB for archival.  Also, did some refactoring on the ServiceManagerBase class, by renaming it to ThreadedService and moving it to ClearCanvas.ImageServer.Common.  Finally, added initial shred support for archival, with a mechanism that polls the database for changes in the configuration for archival and automatically starts/stops the appropriate plugins.