Requirement: #2377
Modified shredhostservice and testapp to use msbuild projects for setting up the output directories.  Note that we're losing some app.config comments that were previously included with this change, because the app config combination utility does not preserve comments.  Still need to do something about this to support web projects.