Requirement: #2389 Add support for separators to action model
Fixed a bug where multiple adjacent separators might appear.  Because the action models are built dynamically based on what tools are available in a given context, you cannot make reliable guarantees about the placement of separators.  If none of the actions between 2 separators actually exist for a given context, those separators will appear in the menu adjacent to one another.

This fix is a bit of a kludge to remove duplicate adjacent separators.  The logic is applied in the view layer by the ToolStripBuilder class, 'cause it was just easier that way.