Ticket: #1910 Ticket: #1911
Renamed the StorageFilesystem table to FilesystemStudyStorage, to match the ArchiveStudyStorage table name.  Added a foreign key to the ServerTransferSyntax table from both the ArchiveStudyStorage and FilesystemStudyStorage tables, so that we know the exact transfer syntax the studies are stored in locally and for archiving.  Added new stored procedure, DeleteFilesystemStudyStorage, to handle purging of data from the cache, and implemented WorkQueue processor for purging.  Introduced running of the rules engine after completion of archiving of a study.  Modified the default "study purge" rule to now occur after archiving.