Ticket: #2478
Removed the ClearCanvas.ImageServer.Services.Shreds assembly.  The classes that were in this assembly are now in each of the individual service plugin dlls.  Had to add some new settings and string resources for this, and also moved around some of the configuration.  The configuration is a bit more straight forward now after the change.