Defect: #2402
Updates to speed up the WorkQueue and StudySearch screens.  We now only get a subset of results when doing a query from the DB, thus speeding up the performance.  Had to add ObjectDataSource objects for binding, and removed a bunch of the manual binding.  This still could use some more cleanup, but it works as a general concept right now.  Some other general code cleanup is included as part of the checkin related to the "Key" property of ServerEntity.