Defect: #2402
Found a problem with the WorkQueue reschedule dialog and fixed it.  Data binding was not updated properly for this dialog.  Also, changed the WorkQueueDataSource to now return WorkQueueSummary objects, which can then be directly displayed in bound fields in the gridview.  This somewhat simplifies things in the WorkQueue screens.  Also, updated the WorkQueueItemListPanel to WorkQueueGridView.  (This was a remnant from way back when we had the accordian for this page instead of a grid view.)