Requirement: #2580
Added new QueueStudyStateEnum table to the database, and put a link in the Study table to the table.  Updated the various insert routines for the WorkQueue and new studies to set this state appropriately, and updated several other places to reset the state value where appropriate.  This is now used to determine which buttons should be used on the GUI.  It also is used to determine if studies are queryable via DICOM.  We now do not return studies who are either scheduled for editing or scheduled for deletion.  Also did some renaming of some of the parameters to stored procedures to match the new naming scheme.