Bug: #2861
Found problem with ReprocessStudyItemProcessor, we had a DB update context opened when doing a bunch of disk access to scan directories.  This was causing a deadlock w/ other processes because the transaction was running for a long time.  Also, when going through this found a number of places where we were doing a query against the study table, but we were only querying with the study instance uid, and not the partition guid, this could cause misleading results if the same study exists on more than one partition.