Requirement: #2873
Introduced a mechanism to make the web GUI server rules page entirely "plugable" for adding new rules.  Simply defining a new rule type, and then defining properly a sample rule in a plugin will cause the rule type to be displayed and be properly added and editable in the web GUI.  No more changes will be needed if we add new rule types.  Also discovered that we were not able to include different tier 1 retention rules, due to some bugs in the code.  We now have a sample rule and can add/edit these rules.