Requirement: #2901
Added new property to FindScuBase, ReuseAssociation.  When set to true, the association is left open, and subsequent calls to Find() on the FindScuBase will reuse the association.  Note also, a new CloseAssociation() method was added to FindScuBase in order to close the association when complete.

Also found a bug in StudyQueryIod, where we were not asking for the tag Number of Study Related Instances.  This was causing the migration tool to do extra queries at the series level.  Finally, founda bug where our check on the count() on the destination device for a study was not working properly either.