Bug: #2899, #2682
Fixed defect where study xml had the wrong filename embedded in it.  There was a setting when creating the xml file to set whether the filename was embedded in the study xml.  This was not implemented properly.  Fixed the implementation, and made it so the server does not store the filename.  Also fixed minor problem with reinventories with previous checkin, where we were logging that a readable storage location was not found for a study with every study we detected for reinventory.  We should no longer log this now...