Bug: #2770
Fixed WorkQueue screen so searches based on patient ID work.  The problem was caused by the WebQueryWorkQueue stored procedure and how it was getting counts of resultant records when query parameters were added.  Required joins were not being added when doing the count.  Found the same problem in WebQueryRestoreQueue and WebQueryArchiveQueue and fixed the problem there too.