Bug: #2979
Fixed problem with decompression of jpeg images.  The problem arose from images having their BitsStored set to 12, but the actual JPEG data being encoded as 16 bit data.  The codecs were clipping the bottom 4 bits off in this case.  Added some code to check the bit depth encoded in the jpeg data, and use that to select the proper codec to use when doing decompression.

Also as part of this, introduced a new tab to the ClearCanvas.Dicom.Samples project for doing compression and /or decompression to single DICOM objects, also added new sample plugin assembly that creates a valid plugin for all the default codecs.